+++ src/eve.helper.spec.ts
@@ -22,6 +22,9 @@ describe('URL creators', () => {
 
         [EVE.getCharacterUrl(5), 'https://esi.evetech.net/v4/characters/5/'],
         [EVE.getCharacterAttributesUrl(5), 'https://esi.evetech.net/v1/characters/5/attributes/'],
+        [EVE.getCharacterAssetsUrl(5), 'https://esi.evetech.net/v3/characters/5/assets/'],
+        [EVE.getCharacterAssetsLocationsUrl(5), 'https://esi.evetech.net/v2/characters/5/assets/locations/'],
+        [EVE.getCharacterAssetsNamesUrl(5), 'https://esi.evetech.net/v1/characters/5/assets/names/'],
         [EVE.getCharacterBlueprintsUrl(5), 'https://esi.evetech.net/v2/characters/5/blueprints/'],
         [EVE.getCharacterShipUrl(5), 'https://esi.evetech.net/v1/characters/5/ship/'],
         [EVE.getCharacterSkillQueueUrl(5), 'https://esi.evetech.net/v2/characters/5/skillqueue/'],